# Reduce function calls in hot path