@mosabua 2024-01-10
Related .. does this still apply and fix the issue for Java 21 .. since that is now the required runtime.